+++ frontend/src/component/main-config-view/main-config-view.tsx
@@ -60,8 +60,8 @@ const CONFIG_REVERSE_PROXY_FIELDS = [
 
 const CONFIG_REVERSE_PROXY_STREAM_FIELDS = [
     {name: 'retry', label: 'LABEL.RECONNECT', fieldType: FormFieldType.CHECK},
-    {name: 'throttle_kbps', label: 'LABEL.THROTTLE_KBPS', fieldType: FormFieldType.TEXT},
-    {name: 'grace_period_millis', label: 'LABEL.GRACE_PERIOD', fieldType: FormFieldType.TEXT},
+    {name: 'throttle', label: 'LABEL.THROTTLE', hint: 'HINT.CONFIG.PROXY.THROTTLE', fieldType: FormFieldType.TEXT},
+    {name: 'grace_period_millis', label: 'LABEL.GRACE_PERIOD', fieldType: FormFieldType.NUMBER},
 ];
 
 const CONFIG_REVERSE_PROXY_STREAM_BUFFER_FIELDS = [
@@ -195,7 +195,7 @@ export default function MainConfigView(props: MainConfigViewProps) {
 
             let cfgReverseProxyStreamBuffer: StreamConfig = {
                 retry: reverseProxyStreamConfig.retry,
-                throttle_kpbs: reverseProxyStreamConfig.throttle_kpbs,
+                throttle: reverseProxyStreamConfig.throttle,
                 grace_period_millis: reverseProxyStreamConfig.grace_period_millis,
                 buffer: {
                     enabled: reverseProxyStreamBufferConfig.enabled,